+++ src/backend/marsha/core/serializers/thumbnail.py
@@ -63,14 +63,20 @@ class ThumbnailSerializer(serializers.ModelSerializer):
             The "validated_data" dictionary is returned after modification.
 
         """
+        resource = self.context["request"].resource
+
+        # Set the video field from the payload if there is one and the user is identified
+        # as a proper user object through access rights
+        if self.initial_data.get("video") and not resource:
+            validated_data["video_id"] = self.initial_data.get("video")
+
         if "size" not in self.initial_data:
             raise serializers.ValidationError({"size": ["File size is required"]})
 
-        # user here is a video as it comes from the JWT
-        # It is named "user" by convention in the `rest_framework_simplejwt` dependency we use.
-        resource = self.context["request"].resource
+        # If the request regards a resource, force the video ID on the timed text track
         if not validated_data.get("video_id") and resource:
             validated_data["video_id"] = resource.id
+
         try:
             return super().create(validated_data)
         except IntegrityError as exc: