+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Allow playlist instructor to manage shared live media through API
 - Allow playlist instructor to manage timed text tracks through API
 - Rework Markdown UI
+- Allow playlist admin/instructor to manage thumbnails through API
 
 ## [4.0.0-beta.16] - 2023-02-17